+++ partiql-eval/src/main/kotlin/org/partiql/lang/eval/physical/window/BuiltInWindowFunction.kt
@@ -9,3 +9,11 @@ internal fun createBuiltinWindowFunction(name: String) =
         "lead" -> Lead()
         else -> error("Window function $name has not been implemented")
     }
+
+@ExperimentalWindowFunctions
+internal fun createBuiltinWindowFunctionAsync(name: String) =
+    when (name) {
+        "lag" -> LagAsync()
+        "lead" -> LeadAsync()
+        else -> error("Window function $name has not been implemented")
+    }