+++ partiql-lang/src/testFixtures/kotlin/org/partiql/lang/eval/evaluatortestframework/PipelineFactory.kt
@@ -4,8 +4,8 @@ import org.partiql.lang.eval.EvaluationSession
 
 /**
  * The implementation of this interface is passed to the constructor of [PipelineEvaluatorTestAdapter].  Determines
- * which pipeline (either [org.partiql.lang.CompilerPipeline] or [org.partiql.lang.compiler.PartiQLCompilerPipeline]) will be
- * tested.
+ * which pipeline (either [org.partiql.lang.CompilerPipeline], [org.partiql.lang.CompilerPipelineAsync],
+ * [org.partiql.lang.compiler.PartiQLCompilerPipeline]) will be tested.
  */
 interface PipelineFactory {
     val pipelineName: String